requirements.txt
langchain 
wikipedia
langchain-community 
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
pydantic 
fastapi
uvicorn[standard]
duckduckgo-search
streamlit
requests

